refactor: re-structure plugin modules
