+++ packages/gunshi/src/plugin/core.test-d.ts
@@ -1,6 +1,6 @@
 import { expectTypeOf, test } from 'vitest'
 
-import { PluginExtension } from './plugin.ts'
+import { PluginExtension } from './core.ts'
 
 test('PluginExtension', () => {
   type T1 = PluginExtension<{ foo: number }>
