+++ packages/gunshi/src/types.ts
@@ -6,7 +6,7 @@
 import { ARG_PREFIX, BUILT_IN_KEY_SEPARATOR, BUILT_IN_PREFIX } from './constants.ts'
 
 import type { Args, ArgToken, ArgValues } from 'args-tokens'
-import type { Plugin } from './plugin.ts'
+import type { Plugin } from './plugin/core.ts'
 
 export type Awaitable<T> = T | Promise<T>
