+++ packages/plugin/README.md
@@ -1,0 +1,7 @@
+# `@gunshi/plugin`
+
+utilities for gunshi plugin
+
+## ©️ License
+
+[MIT](http://opensource.org/licenses/MIT)
