+++ packages/plugin/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * @author kazuya kawaguchi (a.k.a. kazupon)
+ * @license MIT
+ */
+
+export * from 'gunshi/plugin'
